fixed issues, setup commonFunc file
