+++ e2e/src/pages/HomePage.ts
@@ -1,4 +1,4 @@
-import { Page } from '@playwright/test';
+import { Page, expect } from '@playwright/test';
 import BasePage from '../base/BasePage';
 import PostsPreview from '../components/PostsPreview';
 import Footer from '../components/Footer';
@@ -11,5 +11,10 @@ export default class HomePage extends BasePage {
         super(page, 'Home Page', '');
         this.postsPreview = new PostsPreview(page);
         this.footer = new Footer(page);
+    }
+
+    public async isVisible(): Promise<void> {
+        await super.open();
+        expect(this.page.url()).toContain(await super.getPageUrl());
     }
 }
