+++ e2e/tests/postsPreviewCheck.spec.ts
@@ -1,19 +1,19 @@
 import { test, expect } from '../src/FixtureConfigs';
+import { elementsAreVisible, getAllLinks, verifyLinksResponse } from '../src/commonFunctions';
 
 test.describe('Preview Post Validations', () => {
     test('Ensure the home page displays preview posts with correctly functioning links', {
         tag: ['@regression', '@smoke']
-    }, async ({ page, homePage }) => {
+    }, async ({ homePage }) => {
         const amoutOfPosts = 15;
 
-        await homePage.open();
-        expect(page.url()).toContain(await homePage.getPageUrl());
+        await homePage.isVisible();
 
-        await homePage.postsPreview.elementsAreVisible(homePage.postsPreview.allPosts);
+        await elementsAreVisible(homePage.postsPreview.allPosts);
         expect(await homePage.postsPreview.getAllPosts()).toHaveLength(amoutOfPosts);
         expect(await homePage.postsPreview.getAllReadMoreLinks()).toHaveLength(amoutOfPosts);
 
-        const linksUrls = await homePage.postsPreview.getAllLinks(homePage.postsPreview.allReadMoreLinks);
-        await homePage.postsPreview.verifyLinksResponse(linksUrls);
+        const linksUrls = await getAllLinks(homePage.postsPreview.allReadMoreLinks, await homePage.getPageUrl());
+        await verifyLinksResponse(linksUrls);
     });
 });
